requirements.txt
# Piyukonek - Python Dependencies
# Install: pip install -r requirements.txt

Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Mail>=0.9.1
PyMySQL>=1.0.0
Werkzeug>=2.0.0
python-dotenv>=1.0.0
reportlab>=3.6.0
itsdangerous>=2.0.0
requests>=2.28.0
openai>=0.27.0
gunicorn>=21.0.0
